% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mifa.ci.R
\name{ci.mi.each}
\alias{ci.mi.each}
\title{Find the Fieller interval for each k}
\usage{
ci.mi.each(eig.imp, n.factor, alpha, N, M)
}
\arguments{
\item{M}{}
}
\value{

}
\description{
This function is loaded within ci.mifa.fieller and will compute
Fieller's confidence interval for each of the components of n.factor.
}
