% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci.mifa.fieller.R
\name{ci.mifa.fieller}
\alias{ci.mifa.fieller}
\title{Fieller's confidence intervals for proportion of explained variance}
\usage{
ci.mifa.fieller(cov.mi, n.factor, alpha, N)
}
\arguments{
\item{cov.mi}{List containing the estimated covariance matrix within each
imputed data. One can use the outcome of 'mi.cov' with the name
`cov.mice.imp`.}

\item{n.factor}{a vector indicating the number of factors for which the
confidence interval should be constructed.}

\item{alpha}{the level of significance of the confidence interval.}

\item{N}{sample size}
}
\value{
A matrix contining 100(1-alpha)% confidence inervals for n.factor
factors.
}
\description{
Computes parametric confidence intervals for proportion of explained
variance using Fieller's method.
}
