% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mifa-helpers.R
\name{clean_n_pc}
\alias{clean_n_pc}
\title{Check and clean n_pc argument}
\usage{
clean_n_pc(n_pc, n_cov_vars)
}
\arguments{
\item{n_pc}{User-supplied n_pc argument. An integer vector.}

\item{n_cov_vars}{Number of variables in the covariance matrix}
}
\value{
Cleaned n_pc. An integer vector. Sorted, if applicable.
}
\description{
Check and clean n_pc argument
}
\keyword{internal}
