% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{combine.mi}
\alias{combine.mi}
\title{Combine results from different imputations using Rubin's rules}
\usage{
combine.mi(mi.parm.est, mi.parm.cov)
}
\arguments{
\item{mi.parm.est}{Matrix containing estimated parameters in each imputation
as its rows.}

\item{mi.parm.cov}{List containing the covariance matrix estimated within
each imputation.}
}
\value{
A list:
\describe{
\item{parm.est}{Vector of combined parameter estimates with the same length
as columns in \code{mi.parm.est}.}
\item{parm.cov}{Combined variance-covariance matrix of size n x n, where n
is the number of columns in \code{mi.parm.est}.}
\item{between.cov}{Between imputations variance-covariance matrix of size
n x n, where n is the number of columns in \code{mi.parm.est}.}
}
}
\description{
Applies Rubin's rules to combine estimates and
variance-covariance matrices from different imputations.
}
\references{
Rubin D. B. Multiple imputation for nonresponse in surveys (2004).
John Wiley & Sons.
}
