% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mifa-ci.R
\name{get_fieller_ci}
\alias{get_fieller_ci}
\title{Find the Fieller interval for each k}
\usage{
get_fieller_ci(eig_imp, n_factors, conf, N, m)
}
\arguments{
\item{eig_imp}{A matrix with each of its columns the eigenvalues of the
estimated covariance matrix for each imputed data.}

\item{n_factors}{Vector indicating number of factors to be used to compute
proportion of explained variance and construct confidence intervals.
The minimum length of this vector is 1 and its maximum length is
the number of columns in data. Defaults to all possible numbers of factors.}

\item{conf}{Confidence level for constructing confidence intervals. The
default is \code{.95}, i.e., a 95\% condifence interval.}

\item{N}{A scalar specifying sample size.}

\item{m}{A scalar specifying number of multiple imputations.}
}
\value{
A vector of length 2, containing the lower and upper bounds of
estimated Fieller's interval.
}
\description{
This function is used by \code{\link[=mifa_ci_fieller]{mifa_ci_fieller()}} to compute Fieller's confidence
intervals for each of the components of n_factors.
}
\keyword{internal}
