% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mice-helpers.R
\name{mice_impute_all_NA}
\alias{mice_impute_all_NA}
\title{Sequential Imputation}
\usage{
mice_impute_all_NA(data, ...)
}
\arguments{
\item{data}{A data frame with missing values coded as \code{NA}.}

\item{...}{Named arguments that will be passed to \code{\link[mice:mice]{mice::mice()}}, such
as
\itemize{
\item \code{m} Number of generated imputations. The default is 5.
\item \code{maxit} Number of iterations for each imputation. The default is 5.
\item \code{method} Method used for imputation.
\item \code{printFlag} Whether to print diagnostic information to console. The
default is TRUE).
See \code{\link[mice:mice]{mice::mice()}} for all options.
}}
}
\value{
A data frame without NAs
}
\description{
Sequential imputations in case that some of the columns are
not imputed due to collinearity, etc.
}
\details{
This function will keep imputing the same data over and over
until all NA's are gone.
}
\keyword{internal}
