% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mifa.R
\name{mifa}
\alias{mifa}
\title{Compute covariance matrix of incomplete data using multiple imputation}
\usage{
mifa(
  data,
  n_factors = 1:ncol(data),
  ci = FALSE,
  conf = 0.95,
  n_boot = 1000,
  ...
)
}
\arguments{
\item{data}{A data frame with missing values coded as \code{NA}.}

\item{n_factors}{Vector indicating number of factors to be used to compute
proportion of explained variance and construct confidence intervals.
The minimum length of this vector is 1 and its maximum length is
the number of columns in data. Defaults to all possible numbers of factors.}

\item{ci}{A character string or character vector indicating which types of
confidence intervals should be constructed. If \code{"bootstrap"}, \code{"fieller"},
or \code{"both"}, the corresponding intervals are computed. If \code{FALSE}
(the default) no confidence intervals will be computed.}

\item{conf}{Confidence level for constructing confidence intervals. The
default is \code{.95}, i.e., a 95\% condifence interval.}

\item{n_boot}{number of bootstrap samples to use for the bootstrap confidence
intervals. The default is 1000.}

\item{...}{Named arguments that will be passed to \code{\link[mice:mice]{mice::mice()}}, such
as
\itemize{
\item \code{m} Number of generated imputations. The default is 5.
\item \code{maxit} Number of iterations for each imputation. The default is 5.
\item \code{method} Method used for imputation.
\item \code{printFlag} Whether to print diagnostic information to console. The
default is TRUE).
See \code{\link[mice:mice]{mice::mice()}} for all options.
}}
}
\value{
A list:
\describe{
\item{cov_combined}{The estimated covariance matrix of the
incomplete data, based on the combined covariance matrices of imputed
data sets.}
\item{cov_imputations}{A list containing the estimated covariance matrixes
for all imputed data sets.}
\item{var_explained}{A data frame containing the estimated proportions of
explained variance for each of specified \code{n.factor} components. Depending o
n \code{ci}, it will also contain the estimated Fieller's (parametric) and/or
bootstrap (nonparametric) confidence interval for the proportion of
variance explained by the different numbers of factors defined by
\code{n_factors}.}
}
}
\description{
Compute covariance matrix of incomplete data using multiple imputation.
For multiple imputation, Multivariate Imputation by Chained Equations
(MICE) from the \link{mice} package is used.
Also provides variance explained by different numbers of factors with
Fieller (parametric) or bootstrap (nonparametric) confidence intervals.
}
\references{
Nassiri, V., Lovik, A., Molenberghs, G., & Verbeke, G. (2018).
On using multiple imputation for exploratory factor analysis of incomplete
data. Behavioral Research Methods 50, 501–517.
\url{https://doi.org/10.3758/s13428-017-1013-4}
}
\seealso{
\code{\link[=mifa_ci_boot]{mifa_ci_boot()}}, \code{\link[=mifa_ci_fieller]{mifa_ci_fieller()}}, \code{\link[mice:mice]{mice::mice()}}
}
