% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mifa-ci.R
\name{mifa_ci_boot}
\alias{mifa_ci_boot}
\title{Compute bootstrap confidence interval for proportion of explained variance}
\usage{
mifa_ci_boot(data, n_factors, conf = 0.95, n_boot = 1000, ...)
}
\arguments{
\item{data}{A data frame with missing values coded as \code{NA}.}

\item{n_factors}{Vector indicating number of factors to be used to compute
proportion of explained variance and construct confidence intervals.
The minimum length of this vector is 1 and its maximum length is
the number of columns in data. Defaults to all possible numbers of factors.}

\item{conf}{Confidence level for constructing confidence intervals. The
default is \code{.95}, i.e., a 95\% condifence interval.}

\item{n_boot}{number of bootstrap samples to use for the bootstrap confidence
intervals. The default is 1000.}

\item{...}{Named arguments that will be passed to \code{\link[mice:mice]{mice::mice()}}, such
as
\itemize{
\item \code{m} Number of generated imputations. The default is 5.
\item \code{maxit} Number of iterations for each imputation. The default is 5.
\item \code{method} Method used for imputation.
\item \code{printFlag} Whether to print diagnostic information to console. The
default is TRUE).
See \code{\link[mice:mice]{mice::mice()}} for all options.
}}
}
\value{
A data frame containing bootstrapped confidence intervals for
variance explained by different number of factors.
}
\description{
Compute bootstrap confidence intervals for the proportion of explained
variance for the covariance of an incomplete data imputed using
multiple imputation.
For multiple imputation, Multivariate Imputation by Chained Equations
(MICE) from the \link{mice} package is used.
}
\details{
This function uses the Shao and Sitter (1996) method to combine multiple
imputation and bootstrapping. The imputations are done using \code{\link[mice:mice]{mice::mice()}}.
}
\references{
Shao, J. & Sitter, R. R. (1996). Bootstrap for imputed survey data.
Journal of the American Statistical Association 91.435 (1996): 1278-1288.
\url{https://dx.doi.org/10.1080/01621459.1996.10476997}
}
\seealso{
\code{\link[=mifa]{mifa()}}, \code{\link[mice:mice]{mice::mice()}}

Other {mifa confidence intervals}: 
\code{\link{mifa_ci_fieller}()}
}
\concept{{mifa confidence intervals}}
