% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mifa-ci.R
\name{mifa_ci_fieller}
\alias{mifa_ci_fieller}
\title{Fieller's confidence intervals for proportion of explained variance}
\usage{
mifa_ci_fieller(cov_imps, n_factors, conf = 0.95, N)
}
\arguments{
\item{cov_imps}{List containing the estimated covariance matrix within
each imputed data. One can use \code{cov_imputations} returned by \code{\link[=mifa]{mifa()}}.}

\item{n_factors}{Vector indicating number of factors to be used to compute
proportion of explained variance and construct confidence intervals.
The minimum length of this vector is 1 and its maximum length is
the number of columns in data. Defaults to all possible numbers of factors.}

\item{conf}{Confidence level for constructing confidence intervals. The
default is \code{.95}, i.e., a 95\% condifence interval.}

\item{N}{A scalar specifying sample size.}
}
\value{
A matrix containing confidence intervals for \code{n_factors} factors
}
\description{
Computes parametric confidence intervals for proportion of explained
variance for given numbers of factors using Fieller's method.
Note that by setting \code{ci = TRUE} in \code{\link[=mifa]{mifa()}}, this confidence
interval can be computed as well.
}
\references{
Fieller, E. C. (1954). Some problems in interval estimation.
Journal of the Royal Statistical Society. Series B (Methodological): 175-185.
}
\seealso{
\code{\link[=mifa]{mifa()}}

Other {mifa confidence intervals}: 
\code{\link{mifa_ci_boot}()}
}
\concept{{mifa confidence intervals}}
