% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mice-helpers.R
\name{stop_constants}
\alias{stop_constants}
\title{A wrapper for a call to mice that throws an error if there are
constants in the data}
\usage{
stop_constants(f)
}
\arguments{
\item{f}{a call to \code{\link[mice:mice]{mice::mice()}}}
}
\value{
The result of calling the provided function or an Error if there is
a constant in one of the columns.
}
\description{
A wrapper for a call to mice that throws an error if there are
constants in the data
}
\keyword{internal}
