% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mifa.ci.R
\name{ci.mifa.bootstrap}
\alias{ci.mifa.bootstrap}
\title{Compute bootstrap confidence interval for proportion of explained variance}
\usage{
ci.mifa.bootstrap(
  data.miss,
  n.factor,
  rep.boot = 1000,
  method.mi,
  maxit.mi,
  alpha
)
}
\arguments{
\item{data.miss}{Dataset with missing values. Missing values should be
shown with NA.}

\item{n.factor}{Vector indicating number of factor should be used to compute
proportion of explained variance or construct confidence intervals.}

\item{rep.boot}{number of bootstrap samples to use for bootstrap confidence
intervals}

\item{method.mi}{the method which should be used for imputation. It can be a
string or a vector of strings of the size equal to number of items. for more
information see R documentations for mice package. The default is 'pmm'.}

\item{maxit.mi}{A scalar giving the number of iterations for each imputation,
for more information see R documentations for mice package. The default is 5.}

\item{alpha}{Significance level for constructing confidence intervals}
}
\value{

}
\description{
This function computes a bootstrap confidence interval for proportion of
explained variance for the covariance of an incomplete data imputed using
MICE.
}
