% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci.mifa.bootstrap.R
\name{ci.mifa.bootstrap}
\alias{ci.mifa.bootstrap}
\title{Compute bootstrap confidence interval for proportion of explained variance}
\usage{
ci.mifa.bootstrap(
  data.miss,
  n.factor,
  rep.boot = 1000,
  method.mi,
  maxit.mi,
  alpha
)
}
\arguments{
\item{data.miss}{a matrix containing the incomplete datset with items as
columns.}

\item{n.factor}{a vector containing the number of factors should be used to
compute proportion of explained variance.}

\item{rep.boot}{the number of sub-samples to construct the bootstrap.
confidence interval.}

\item{method.mi}{the method which should be used for imputation. It can be a
string or a vector of strings of the size equal to number of items.}

\item{maxit.mi}{a scalar giving the number of iterations for each imputation,}

\item{alpha}{confidence level.}
}
\value{

}
\description{
This function computes a bootstrap confidence interval for proportion of
explained variance for the covariance of an incomplete data imputed using
MICE.
Please make sure the 'mice' package is installed. To install it use
`install.packages("mice")`
}
