% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mifa.ci.R
\name{ci.mifa.fieller}
\alias{ci.mifa.fieller}
\title{Fieller's confidence intervals for proportion of explained variance}
\usage{
ci.mifa.fieller(cov.mi, n.factor, alpha, N)
}
\arguments{
\item{cov.mi}{List containing the estimated covariance matrix within each
imputed data. One can use the outcome of 'mifa.cov' with the name
`cov.mice.imp`.}

\item{n.factor}{Vector indicating number of factor should be used to compute
proportion of explained variance or construct confidence intervals.}

\item{alpha}{Significance level for constructing confidence intervals}

\item{N}{sample size}
}
\value{
A matrix containing 100(1-alpha)% confidence intervals for n.factor
factors.
}
\description{
Computes parametric confidence intervals for proportion of explained
variance using Fieller's method.
}
