% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mifa.ci.R
\name{ci.mifa.fieller}
\alias{ci.mifa.fieller}
\title{Fieller's confidence intervals for proportion of explained variance}
\usage{
ci.mifa.fieller(cov.mi, n.factor, alpha, N)
}
\arguments{
\item{cov.mi}{List containing the estimated covariance matrix within
each imputed data. One can use \code{cov.mice.imp} from \code{\link[=mifa.cov]{mifa.cov()}}.}

\item{n.factor}{Vector indicating number of factors to be used to compute
proportion of explained variance or construct confidence intervals.
The minimum length of this vector is 1 and its maximum length is
the number of items.}

\item{alpha}{Significance level for constructing confidence intervals.}

\item{N}{A scalar specifying sample size.}
}
\value{
A matrix containing confidence intervals for \code{n.factor} factors with
a confidence level of \verb{100(1-alpha)\%}.
}
\description{
Computes parametric confidence intervals for proportion of explained
variance for given numbers of factors using Fieller's method.
Note that by setting \code{ci = TRUE} in \code{\link[=mifa.cov]{mifa.cov()}}, this confidence
interval can be computed as well.
}
\references{
Fieller, E. C. (1954). Some problems in interval estimation.
Journal of the Royal Statistical Society. Series B (Methodological): 175-185.
}
\seealso{
\code{\link[=mifa.cov]{mifa.cov()}}

Other {mifa confidence intervals}: 
\code{\link{ci.mifa.bootstrap}()}
}
\concept{{mifa confidence intervals}}
