% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.mi.R
\name{combine.mi}
\alias{combine.mi}
\title{Combine results from different imputations using Rubin's rule}
\usage{
combine.mi(mi.parm.est, mi.parm.cov)
}
\arguments{
\item{mi.parm.est}{Matrix containing estimated parameters in each imputation
as its rows.}

\item{mi.parm.cov}{List containing the covariance matrix estimated within
each imputation.}
}
\value{
A list: `parm.est`, `parm.cov`, `between.cov`
}
\description{
Use Rubin's rule to combine results from different imputations
}
