% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.mi.R
\name{combine.mi}
\alias{combine.mi}
\title{Combine results from different imputations using Rubin's rules}
\usage{
combine.mi(mi.parm.est, mi.parm.cov)
}
\arguments{
\item{mi.parm.est}{Matrix containing estimated parameters in each imputation
as its rows.}

\item{mi.parm.cov}{List containing the covariance matrix estimated within
each imputation.}
}
\value{
A list:
\describe{
\item{parm.est}{Combined estimates.}
\item{parm.cov}{Combined variance-covariance matrix.}
\item{between.cov}{Between imputations variance-covariance matrix.}
}
}
\description{
Applies Rubin's rules to combine estimates and
variance-covariance matrices from different imputations.
}
\references{
Rubin D. B. Multiple imputation for nonresponse in surveys (2004).
John Wiley & Sons.
}
