% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.mi.R
\name{combine.mi}
\alias{combine.mi}
\title{Combine results from different imputations}
\usage{
combine.mi(mi.parm.est, mi.parm.cov)
}
\arguments{
\item{mi.parm.est}{Matrix containing estimated parameters in each imputation
as its rows.}

\item{mi.parm.cov}{List contaning the covariance matrix estimated within
each imputation.}
}
\value{

}
\description{
This function applies Rubin's rule to combine results from different
imputations
}
