% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mice-helpers.R
\name{mice_impute_all_NA}
\alias{mice_impute_all_NA}
\title{Sequential Imputation}
\usage{
mice_impute_all_NA(data, ...)
}
\arguments{
\item{data}{A data frame with missing values coded as \code{NA}.}
}
\value{
A data frame without NAs
}
\description{
Sequential imputations in case that some of the columns are
not imputed due to collinearity, etc.
}
\details{
This function will keep imputing the same data over and over
until all NA's are gone.
}
\keyword{internal}
