% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mice-helpers.R
\name{mice_impute_once}
\alias{mice_impute_once}
\title{Imputes a single dataset using mice}
\usage{
mice_impute_once(data, ...)
}
\arguments{
\item{data}{A data frame with missing values coded as \code{NA}.}
}
\value{
A data frame with missing values imputed.
}
\description{
Imputes a single dataset using mice
}
\keyword{internal}
