% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mifa-ci.R
\name{mifa_ci_boot}
\alias{mifa_ci_boot}
\title{Bootstrap confidence intervals for proportion of explained variance}
\usage{
mifa_ci_boot(data, cov_var, n_factors, conf = 0.95, n_boot = 1000, ...)
}
\arguments{
\item{data}{A data frame with missing values coded as \code{NA}.}

\item{cov_var}{Variables in \code{data} to calculate the covariance
matrix for. Supports (tidy selection)\code{\link[dplyr:select]{dplyr::select()}}. This allows to
select variables that are used for the imputations of missing values, but not
the calculations of the covariance matrix. This is especially useful when
there are categorical predictors that can improve the imputation of the
response variables, but for which covariance cannot be calculated.
By default, all variables in \code{data} are used for both, the imputation and
the covariance matrix. Note: Variables and rows used for the imputation
can be configured using the \code{...}. See also \code{\link[mice:mice]{mice::mice()}}.}

\item{n_factors}{Integer or integer vector indicating number of factors for
which proportion of explained variance (and optionally confidence intervals)
should be indicated. Defaults to all possible numbers of factors.}

\item{conf}{Confidence level for constructing confidence intervals. The
default is \code{.95}, i.e., a 95\% confidence interval.}

\item{n_boot}{number of bootstrap samples to use for the bootstrap confidence
intervals. The default is 1000.}
}
\value{
A data frame containing bootstrapped confidence intervals for
variance explained by different number of factors.
}
\description{
Compute bootstrap confidence intervals for the proportion of explained
variance for the covariance of an incomplete data imputed using
multiple imputation.
For multiple imputation, Multivariate Imputation by Chained Equations
(MICE) from the \link{mice} package is used.
}
\details{
This function uses the Shao and Sitter (1996) method to combine multiple
imputation and bootstrapping. The imputations are done using \code{\link[mice:mice]{mice::mice()}}.
}
\references{
Shao, J. & Sitter, R. R. (1996). Bootstrap for imputed survey data.
Journal of the American Statistical Association 91.435 (1996): 1278-1288.
\url{https://dx.doi.org/10.1080/01621459.1996.10476997}
}
\seealso{
\code{\link[=mifa]{mifa()}}, \code{\link[mice:mice]{mice::mice()}}

Other {mifa confidence intervals}: 
\code{\link{mifa_ci_fieller}()}
}
\concept{{mifa confidence intervals}}
